requirements.txt
speechrecognition
openai-whisper
pyaudio
pycaw
comtypes
urllib3
soundfile
